+++ alacritty/alacritty.xcodeproj/xcuserdata/bbarrows.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "FB2D727B-63AB-42F7-95E1-846056DD5296"
+   type = "1"
+   version = "2.0">
+</Bucket>
